Move util function to reports
